add checkout func

+++ app/Controllers/Http/CartController.js
@@ -3,6 +3,8 @@
 const Cart = use('App/Models/Cart')
 const Product = use('App/Models/Product')
 
+const { validate } = use("Validator");
+
 /** @typedef {import('@adonisjs/framework/src/Request')} Request */
 /** @typedef {import('@adonisjs/framework/src/Response')} Response */
 /** @typedef {import('@adonisjs/framework/src/View')} View */
@@ -21,12 +23,20 @@ class CartController {
      * @param {View} ctx.view
      */
     async index({ request, response, view }) {
-        const cart = await Cart.all(
-            response.status(200).json({
-                message: 'Get all Cart successfully.',
+        try {
+            const cart = await Cart.all()
+            return response.status(200).json({
+                message: 'Get all Cart success.',
                 data: cart
             })
-        )
+        } catch (error) {
+            console.log(error);
+            return response.status(400).json({
+                message: 'Something went wrong!'
+            })
+
+
+        }
     }
 
     /**
@@ -52,37 +62,29 @@ class CartController {
      */
     async store({ request, response }) {
 
-        const { user_id, product_id, product_qty, total } = request.post()
+        try {
+            const rules = {
+                user_id: 'required|number',
+                product_id: 'required|number',
+                product_qty: 'required|number',
+                total: 'required|number',
+            }
+            const validation = await validate(request.all(), rules);
+            if (validation.fails())
+                return response.status(400).json({
+                    message: validation.messages()
+                });
 
-        // Checking if there is already cart for this user 
-        const cartCheck = await Cart.query()
-            .where('user_id', user_id)
-            .getCount()
 
-        if (cartCheck < 1) {
-            // if there's none make one
-            const cart = new Cart()
-            cart.product_id = product_id
-            cart.product_qty = product_qty
-            cart.price = total
-            cart.total = total * product_qty
-            cart.user_id = user_id
+            const { user_id, product_id, product_qty, total } = request.post()
 
-            await cart.save()
-            response.status(200).json({
-                message: 'Cart added successfully.',
-                data: cart
-            })
-
-        } else {
-            // if there is one then check if the cart has the product
-            const cart = await Cart.query()
+            // Checking if there is already cart for this user 
+            const cartCheck = await Cart.query()
                 .where('user_id', user_id)
-                .where('product_id', product_id)
                 .getCount()
 
-            if (cart < 1) {
-                // if the cart doesnt has the product add it 
+            if (cartCheck < 1) {
+                // if there's none make one
                 const cart = new Cart()
                 cart.product_id = product_id
                 cart.product_qty = product_qty
@@ -91,46 +93,77 @@ class CartController {
                 cart.user_id = user_id
 
                 await cart.save()
-
-                const cartPrice = await Cart.query()
-                    .where('user_id', user_id)
-                    .sum('total as total')
-
-                response.status(200).json({
-                    message: 'Checkout added successfully.',
-                    data: cart,
-                    total: cartPrice[0].total
+                return response.status(201).json({
+                    message: 'Cart added successfully.',
+                    data: cart
                 })
 
             } else {
-                // if the cart has the product then add the qty of the product
-                const cartCheck = await Cart.query()
-                    .select('id')
+                // if there is one then check if the cart has the product
+                const cart = await Cart.query()
                     .where('user_id', user_id)
                     .where('product_id', product_id)
-                    .fetch()
+                    .getCount()
 
-                const idCart = cartCheck.rows[0].id
+                if (cart < 1) {
+                    // if the cart doesnt has the product add it 
+                    const cart = new Cart()
+                    cart.product_id = product_id
+                    cart.product_qty = product_qty
+                    cart.price = total
+                    cart.total = total * product_qty
+                    cart.user_id = user_id
 
-                const cart = await Cart.find(idCart)
-                cart.product_qty = cart.product_qty + 1
-                cart.total = cart.price * cart.product_qty
+                    await cart.save()
 
-                await cart.save()
+                    const cartPrice = await Cart.query()
+                        .where('user_id', user_id)
+                        .sum('total as total')
 
-                const cartPrice = await Cart.query()
-                    .where('user_id', user_id)
-                    .sum('total as total')
+                    return response.status(201).json({
+                        message: 'Cart added successfully.',
+                        data: cart,
+                        total: cartPrice[0].total
+                    })
 
-                response.status(200).json({
-                    message: 'Checkout added successfully.',
-                    data: cart,
-                    total: cartPrice[0].total
-                })
+                } else {
+                    // if the cart has the product then add the qty of the product
+                    const cartCheck = await Cart.query()
+                        .select('id')
+                        .where('user_id', user_id)
+                        .where('product_id', product_id)
+                        .fetch()
+
+                    const idCart = cartCheck.rows[0].id
+
+                    const cart = await Cart.find(idCart)
+                    cart.product_qty = cart.product_qty + 1
+                    cart.total = cart.price * cart.product_qty
+
+                    await cart.save()
+
+                    const cartPrice = await Cart.query()
+                        .where('user_id', user_id)
+                        .sum('total as total')
+
+                    return response.status(201).json({
+                        message: 'Cart added successfully.',
+                        data: cart,
+                        total: cartPrice[0].total
+                    })
+
+                }
 
             }
 
+        } catch (error) {
+            return response.status(400).json({
+                message: 'Something went wrong!'
+            })
+
         }
+
+
 
     }
 
@@ -145,20 +178,35 @@ class CartController {
      */
     async show({ params, request, response, view }) {
 
-        const product = await Product.query()
-            .leftJoin('carts', 'products.id', 'carts.product_id')
-            .where('carts.user_id', params.id)
-            .fetch()
+        try {
+            const product = await Product.query()
+                .leftJoin('carts', 'products.id', 'carts.product_id')
+                .where('carts.user_id', params.id)
+                .fetch()
 
-        const cartPrice = await Cart.query()
-            .where('user_id', params.id)
-            .sum('total as total')
+            const cartPrice = await Cart.query()
+                .where('user_id', params.id)
+                .sum('total as total')
+                
+            const productWeight = await Product.query()
+                .leftJoin('carts', 'products.id', 'carts.product_id')
+                .where('carts.user_id', params.id)
+                .sum('weight as weight')
 
-        response.status(200).json({
-            status: 1,
-            data: product,
-            total: cartPrice[0].total
-        })
+            return response.status(200).json({
+                message: 'Get cart for this user success.',
+                data: product,
+                total: cartPrice[0].total,
+                weight: productWeight[0].weight
+            })
+
+        } catch (error) {
+            return response.status(400).json({
+                message: 'Something went wrong!'
+            })
+
+        }
+
     }
 
     /**
@@ -183,41 +231,51 @@ class CartController {
      */
     async update({ params, request, response }) {
 
-        const { qty } = request.post()
+        try {
+            const { qty } = request.post()
 
-        const idCart = parseInt(params.id)
+            const idCart = parseInt(params.id)
 
-        // const idProduct = parseInt(params.product)
+            // const idProduct = parseInt(params.product)
 
-        const cart = await Cart.find(idCart)
-        if (cart == null) {
-            return response.status(404).json({
-                message: 'Data not found.',
+            const cart = await Cart.find(idCart)
+            if (cart == null) {
+                return response.status(404).json({
+                    message: 'Product not found.',
+                })
+            }
+            const total = cart.price * qty
+            const idUser = cart.user_id
+            const idProduct = cart.product_id
+            cart.product_qty = qty
+            cart.total = total
+
+            await cart.save()
+
+            const product = await Product.query()
+                .leftJoin('carts', 'products.id', 'carts.product_id')
+                .where('carts.product_id', idProduct)
+                .where('carts.user_id', idUser)
+                .fetch()
+            const cartPrice = await Cart.query()
+                .where('user_id', idUser)
+                .sum('total as total')
+
+            return response.status(200).json({
+                message: 'Cart updated successfully.',
+                data: product.rows[0],
+                total: cartPrice[0].total
             })
+
+        } catch (error) {
+            return response.status(400).json({
+                message: 'something went wrong!'
+            })
+
         }
-        const total = cart.price * qty
-        const idUser = cart.user_id
-        const idProduct = cart.product_id
-        cart.product_qty = qty
-        cart.total = total
 
-        await cart.save()
 
-        const product = await Product.query()
-            .leftJoin('carts', 'products.id', 'carts.product_id')
-            .where('carts.product_id', idProduct)
-            .where('carts.user_id', idUser)
-            .fetch()
-        const cartPrice = await Cart.query()
-            .where('user_id', idUser)
-            .sum('total as total')
-
-        response.status(200).json({
-            message: 'Checkout updated successfully.',
-            data: product.rows[0],
-            total: cartPrice[0].total
-        })
-        // const checkout = await Checkout.query()
+        // const Cart = await Cart.query()
         // .select('id')
         // .where('user_id', params.id)
         // .where('status', 0)
@@ -234,23 +292,36 @@ class CartController {
      * @param {Response} ctx.response
      */
     async destroy({ params, request, response }) {
-        const cart = await Cart.find(params.id)
-        const idUser = cart.user_id
-        await cart.delete()
 
-        const product = await Product.query()
-            .leftJoin('carts', 'products.id', 'carts.product_id')
-            .where('carts.user_id', idUser)
-            .fetch()
-        const cartPrice = await Cart.query()
-            .where('user_id', idUser)
-            .sum('total as total')
+        try {
+            const cart = await Cart.find(params.id)
+            const idUser = cart.user_id
+            if (cart == null)
+                return response.status(404).json({
+                    message: 'product not found.'
+                })
+            await cart.delete()
 
-        response.status(200).json({
-            message: 'Checkout deleted successfully.',
-            data: product,
-            total: cartPrice[0].total
-        })
+            const product = await Product.query()
+                .leftJoin('carts', 'products.id', 'carts.product_id')
+                .where('carts.user_id', idUser)
+                .fetch()
+            const cartPrice = await Cart.query()
+                .where('user_id', idUser)
+                .sum('total as total')
+
+            return response.status(200).json({
+                message: 'Cart deleted successfully.',
+                data: product,
+                total: cartPrice[0].total
+            })
+
+        } catch (error) {
+            return response.status(400).json({
+                message: 'something went wrong!'
+            })
+
+        }
 
     }
 }
